+++ packages/apollos-data-connector-postgres/src/index.js
@@ -7,6 +7,8 @@ export {
 } from './postgres';
 
 export * as Comment from './comments';
+export * as ContentItem from './content-items';
+export * as ContentItemsConnection from './content-items-connections';
 export * as UserFlag from './user-flags';
 export * as UserLike from './user-likes';
 export * as Follow from './follows';
@@ -14,5 +16,6 @@ export * as Person from './people';
 export * as Campus from './campus';
 export * as Notification from './notifications';
 export * as NotificationPreference from './notification-preferences';
+export * as Media from './media';
 
 export createMigrationRunner from './postgres/performMigrations';