+++ packages/apollos-data-connector-postgres/src/media/migrations/001_create_media.js
@@ -1,0 +1,21 @@
+import { DataTypes } from 'sequelize';
+
+async function up({ context: queryInterface }) {
+  await queryInterface.createTable('media', {
+    type: DataTypes.ENUM(['IMAGE', 'VIDEO', 'AUDIO']),
+    url: DataTypes.STRING,
+    nodeId: DataTypes.UUID,
+    nodeType: DataTypes.STRING,
+  });
+  await queryInterface.addIndex('media', ['originId', 'originType'], {
+    unique: true,
+  });
+}
+
+async function down({ context: queryInterface }) {
+  await queryInterface.dropTable('media');
+}
+
+const name = '001-create-media';
+
+module.exports = { up, down, name, order: 4 };